requirements.txt
requests
tweepy
praw
googletrans==4.0.0-rc1
python-dotenv
urllib3
